fix: resolve compiler warnings in benchmarks and examples - Fix printf format specifiers for uint64_t (%d -> %lu) - Fix signed/unsigned comparison warnings with type casts - Add [[fallthrough]] attribute for intentional switch fallthrough - Remove unused variables and functions - Replace unsafe tmpnam() with safe temporary filenames

+++ benchmarks/src/random_read.cpp
@@ -46,7 +46,7 @@ static void BM_stdio_RandomRead(benchmark::State &state) {
         }
 
         auto actual_file_size = ftell(file);
-        if (actual_file_size != file_size) {
+        if (static_cast<size_t>(actual_file_size) != file_size) {
             state.SkipWithError("wrong file_size: " + std::to_string(actual_file_size) +
                                 " != " + std::to_string(file_size));
         }
